timesheet entry routes are now limited to the individual user

+++ app/routes/timesheets-entries.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const timesheetEntryService = require('./../services/timesheet-entry-service');
+var stormpath = require('express-stormpath');
 
 const router = express.Router();
 
-router.get('/', (req, res) => {
+router.get('/', stormpath.getUser, (req, res) => {
     timesheetEntryService.find(req).then((entries) => {
         res.send(entries);
     }).catch((reason) => {
@@ -12,7 +13,7 @@ router.get('/', (req, res) => {
     });
 });
 
-router.post('/', (req, res) => {
+router.post('/', stormpath.getUser, (req, res) => {
     req.checkBody('employerOffice.id', 'Employer\'s office identifier is required').notEmpty();
     req.checkBody('start', 'Start is required').notEmpty();
     req.checkBody('end', 'End is required').notEmpty();
@@ -25,10 +26,12 @@ router.post('/', (req, res) => {
         return;
     }
 
+    var userId = getUserId(req);
     timesheetEntryService.create({
         employerOffice: {
             id: req.body.employerOffice.id
         },
+        userId: userId,
         start: req.body.start,
         end: req.body.end,
         break: req.body.break,
@@ -41,8 +44,17 @@ router.post('/', (req, res) => {
     });
 });
 
-router.get('/:id', (req, res) => {
+router.get('/:id', stormpath.getUser, (req, res) => {
     timesheetEntryService.findById(req.params.id, req).then((entry) => {
+        if(entry === null) {
+            res.sendStatus(404);
+            return;
+        }
+        var userId = getUserId(req);
+        if(entry.userId !== userId) {
+            res.sendStatus(403);
+            return;
+        }
         res.send(entry);
     }).catch((reason) => {
         console.log(reason);
@@ -50,13 +62,28 @@ router.get('/:id', (req, res) => {
     });
 });
 
-router.delete('/:id', (req, res) => {
-    timesheetEntryService.remove(req.params.id, req).then(() => {
-        res.sendStatus(204);
+router.delete('/:id', stormpath.getUser, (req, res) => {
+    timesheetEntryService.findById(req.params.id, req).then((entry) => {
+        if(entry === null) {
+            res.sendStatus(404);
+            return;
+        }
+        var userId = getUserId(req)
+        if(entry.userId !== userId) {
+            res.sendStatus(403);
+            return;
+        }
+        timesheetEntryService.remove(req.params.id, req).then(() => {
+            res.sendStatus(204);
+        });
     }).catch((reason) => {
         console.log(reason);
         res.sendStatus(500);
     });
 });
+
+function getUserId(req) {
+    return req.user.href.substr(req.user.href.lastIndexOf('/') + 1);
+}
 
 module.exports = router;
